@@ -1,93 +1,53 @@
-import {
-  withAndroidManifest,
-  createRunOncePlugin,
-  ExportedConfigWithProps,
-  ExportedConfig,
-} from '@expo/config-plugins';
-
-// eslint-disable-next-line import/no-commonjs, @typescript-eslint/no-var-requires
-const pkg = require('../../package.json');
+import { ExportedConfig } from '@expo/config-plugins';
+import { withBuildProperties } from 'expo-build-properties';
 
 /**
- * @type {import('./types').ManifestQueries}
- * what we are trying to add:
- * <queries>
-    <package android:name="com.facebook.katana"/>
-    <package android:name="com.instagram.android"/>
-    <package android:name="com.twitter.android"/>
-    <package android:name="com.zhiliaoapp.musically"/>
-    <intent></intent>
-      <action android:name="android.intent.action.VIEW"/>
-      <category android:name="android.intent.category.BROWSABLE"/>
-      <data android:scheme="https"/>
-    </intent>
-  </queries>
+ * Handles for edge case when LSApplicationQueriesSchemes is an object or undefined.
  */
-
-/**
- * @param {import('@expo/config-plugins').ExportedConfig} config
- */
-const withAndroidManifestService = (config: ExportedConfig, props: WithSocialShareProps) => {
-  return withAndroidManifest(config, (config: ExportedConfigWithProps) => {
-    config.modResults.manifest = {
-      ...config.modResults.manifest,
-      queries: {
-        package: props?.android?.map((social) => ({
-          $: {
-            'android:name': social,
-          },
-        })),
-        intent: [
-          {
-            action: {
-              $: {
-                'android:name': 'android.intent.action.VIEW',
-              },
-            },
-            category: {
-              $: {
-                'android:name': 'android.intent.category.BROWSABLE',
-              },
-            },
-            data: {
-              $: {
-                'android:scheme': 'https',
-              },
-            },
-          },
-        ],
-      },
-    };
-
-    return config;
-  });
+const getIOSQuerySchemes = (config: ExportedConfig): string[] => {
+  return Array.isArray(config.ios?.infoPlist?.LSApplicationQueriesSchemes)
+    ? config.ios?.infoPlist?.LSApplicationQueriesSchemes ?? []
+    : [];
 };
 
-const withInfoPlist = (config: ExportedConfig, props: WithSocialShareProps) => {
-  return {
-    ...config,
-    ios: {
-      ...config.ios,
-      infoPlist: {
-        ...config.ios?.infoPlist,
-        LSApplicationQueriesSchemes: config.ios?.infoPlist?.LSApplicationQueriesSchemes
-          ? [...config.ios.infoPlist.LSApplicationQueriesSchemes, ...props.ios]
-          : props.ios,
+export default (
+  config: ExportedConfig,
+  props: {
+    enableBase64ShareAndroid?: boolean;
+    android?: string[];
+    ios?: string[];
+  },
+) => {
+  return withBuildProperties(
+    {
+      ...config,
+      android: {
+        ...config.android,
+        ...(props.enableBase64ShareAndroid
+          ? {
+              permissions: [
+                ...new Set([
+                  ...(config.android?.permissions ?? []),
+                  'android.permission.WRITE_EXTERNAL_STORAGE',
+                ]),
+              ],
+            }
+          : {}),
+      },
+      ios: {
+        ...config.ios,
+        infoPlist: {
+          ...config.ios?.infoPlist,
+          LSApplicationQueriesSchemes: [...getIOSQuerySchemes(config), ...(props?.ios ?? [])],
+        },
       },
     },
-  };
+    {
+      android: {
+        manifestQueries: {
+          package: props.android ?? [],
+        },
+      },
+    },
+  );
 };
-
-type WithSocialShareProps = {
-  ios: string[];
-  android: string[];
-};
-
-function withSocialShare(config: ExportedConfig, props: WithSocialShareProps) {
-  config = withAndroidManifestService(config, props); // Android
-  config = withInfoPlist(config, props); // iOS
-  return config;
-}
-
-// eslint-disable-next-line import/no-commonjs
-module.exports = createRunOncePlugin(withSocialShare, pkg.name, pkg.version);
